+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
       MYSQL_CHARACTER_SET_SERVER: utf8
       MYSQL_CHARACTER_SET_CLIENT: utf8
     volumes:
-      - ./app/database/init.sql:/docker-entrypoint-initdb.d/init.sql
+      - ./init.sql:/docker-entrypoint-initdb.d/init.sql
     ports:
       - "3306:3306"
     networks:
